+++ commerce/delivery/src/main/java/ru/yandex/practicum/delivery/service/DeliveryService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.delivery.service;
 import ru.yandex.practicum.interactionapi.dto.DeliveryDto;
 import ru.yandex.practicum.interactionapi.dto.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -15,5 +16,5 @@ public interface DeliveryService {
 
     void deliveryFailed(UUID deliveryId);
 
-    Double deliveryCost(OrderDto orderDto);
+    BigDecimal deliveryCost(OrderDto orderDto);
 }
