+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interactionapi/dto/PaymentDto.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.interactionapi.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -12,7 +13,7 @@ import java.util.UUID;
 public class PaymentDto {
     @NotNull
     private UUID paymentId;
-    private Double totalPayment;
-    private Double deliveryTotal;
-    private Double feeTotal;
+    private BigDecimal totalPayment;
+    private BigDecimal deliveryTotal;
+    private BigDecimal feeTotal;
 }
